feat: Ssgoi 기본 설정
