+++ package.json
@@ -9,6 +9,7 @@
     "lint": "eslint"
   },
   "dependencies": {
+    "@ssgoi/react": "^4.2.1",
     "@stomp/stompjs": "^7.2.1",
     "@tanstack/react-query": "^5.90.12",
     "axios": "^1.13.2",
