+++ src/app/onboarding/page.tsx
@@ -1,11 +1,12 @@
+import { Suspense } from "react";
 import OnboardingWizard from "@/src/features/onboarding/ui/OnboardingWizard";
 
 export default function OnboardingPage() {
   return (
-    <>
+    <Suspense fallback={null}>
       <div>
         <OnboardingWizard />
       </div>
-    </>
+    </Suspense>
   );
 }
