+++ src/app/template.tsx
@@ -1,0 +1,10 @@
+import type { ReactNode } from "react";
+import PageTransition from "./page-transition";
+
+export default function Template({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return <PageTransition>{children}</PageTransition>;
+}
